+++ pyanalyze/annotations.py
@@ -63,6 +63,7 @@ from .extensions import (
     NoReturnGuard,
     ParameterTypeGuard,
     TypeGuard,
+    deprecated,
 )
 from .find_unused import used
 from .functions import FunctionDefNode
@@ -949,6 +950,12 @@ class Pep655Value(Value):
     value: Value
 
 
+@dataclass
+class DecoratorValue(Value):
+    decorator: object
+    args: Tuple[Value, ...]
+
+
 class _Visitor(ast.NodeVisitor):
     def __init__(self, ctx: Context) -> None:
         self.ctx = ctx
@@ -1100,6 +1107,14 @@ class _Visitor(ast.NodeVisitor):
                 return None
             tv = ParamSpec(name_val.val)
             return TypeVarValue(tv, is_paramspec=True)
+        elif is_typing_name(func.val, "deprecated") or func.val is deprecated:
+            if node.keywords:
+                self.ctx.show_error(
+                    "deprecated() does not accept keyword arguments", node=node
+                )
+                return None
+            arg_values = tuple(self.visit(arg) for arg in node.args)
+            return DecoratorValue(deprecated, arg_values)
         elif isinstance(func.val, type):
             if func.val is object:
                 return AnyValue(AnySource.inference)