+++ pyanalyze/typeshed.py
@@ -39,13 +39,19 @@ from .analysis_lib import is_positional_only_arg_name
 from .annotations import (
     Context,
     make_type_var_value,
+    DecoratorValue,
     Pep655Value,
     SyntheticEvaluator,
     type_from_value,
     value_from_ast,
 )
 from .error_code import ErrorCode
-from .extensions import evaluated, overload, real_overload
+from .extensions import (
+    evaluated,
+    overload,
+    real_overload,
+    deprecated as deprecated_decorator,
+)
 from .node_visitor import Failure
 from .options import Options, PathSequenceOption
 from .safe import all_of_type, hasattr_static, is_typing_name, safe_isinstance
@@ -63,6 +69,9 @@ from .value import (
     AnyValue,
     CallableValue,
     CanAssignContext,
+    DeprecatedExtension,
+    Extension,
+    annotate_value,
     extract_typevars,
     GenericValue,
     KnownValue,
@@ -795,6 +804,7 @@ class TypeshedFinder:
         allow_call: bool = False,
     ) -> Optional[Signature]:
         is_classmethod = is_staticmethod = is_evaluated = False
+        deprecated = None
         for decorator_ast in node.decorator_list:
             decorator = self._parse_expr(decorator_ast, mod)
             if (
@@ -814,8 +824,14 @@ class TypeshedFinder:
             elif decorator == KnownValue(evaluated):
                 is_evaluated = True
                 continue
-            # might be @overload or something else we don't recognize
-            return None
+            elif (
+                isinstance(decorator, DecoratorValue)
+                and decorator.decorator is deprecated_decorator
+            ):
+                arg = decorator.args[0]
+                if isinstance(arg, KnownValue) and isinstance(arg.val, str):
+                    deprecated = arg.val
+            # something we don't recognize; ignore it
         if node.returns is None:
             return_value = AnyValue(AnySource.unannotated)
         else:
@@ -876,6 +892,7 @@ class TypeshedFinder:
             else return_value,
             allow_call=allow_call,
             evaluator=evaluator,
+            deprecated=deprecated,
         )
 
     def _parse_param_list(
@@ -982,10 +999,27 @@ class TypeshedFinder:
         ctx = _AnnotationContext(finder=self, module=module)
         return value_from_ast(info.ast.value, ctx=ctx)
 
+    def _extract_metadata(self, module: str, node: ast.ClassDef) -> Sequence[Extension]:
+        metadata = []
+        for decorator in node.decorator_list:
+            decorator_val = self._parse_expr(decorator, module)
+            if (
+                isinstance(decorator_val, DecoratorValue)
+                and decorator_val.decorator is deprecated_decorator
+            ):
+                arg = decorator_val.args[0]
+                if isinstance(arg, KnownValue) and isinstance(arg.val, str):
+                    metadata.append(DeprecatedExtension(arg.val))
+        return metadata
+
     def make_synthetic_type(self, module: str, info: typeshed_client.NameInfo) -> Value:
         fq_name = f"{module}.{info.name}"
         bases = self._get_bases_from_info(info, module, fq_name)
         typ = TypedValue(fq_name)
+        if isinstance(info.ast, ast.ClassDef):
+            metadata = self._extract_metadata(module, info.ast)
+        else:
+            metadata = []
         if bases is not None:
             if any(
                 (isinstance(base, KnownValue) and is_typing_name(base.val, "TypedDict"))
@@ -993,7 +1027,10 @@ class TypeshedFinder:
                 for base in bases
             ):
                 typ = self._make_typeddict(module, info, bases)
-        return SubclassValue(typ, exactly=True)
+        val = SubclassValue(typ, exactly=True)
+        if metadata:
+            return annotate_value(val, metadata)
+        return val
 
     def _make_typeddict(
         self, module: str, info: typeshed_client.NameInfo, bases: Sequence[Value]